pyproject.toml
[project]
name = "chaidocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-genai>=1.14.0",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.8",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
]
